+++ src/app/components/userInformationForm/insuranceDetails/InsuranceDetails.jsx
@@ -2,13 +2,26 @@ import { Typography, Form, Select, Row, Col } from "antd";
 import React, { useState } from "react";
 import InsuranceAttestaion from "../common/insuranceAttestation/InsuranceAttestation.jsx";
 import InsuranceInformation from "../common/insuranceInformation/InsuranceInformation.jsx";
-import {
-  insuranceCompanySelectAttributes,
-  insuranceNumberInputAttributes,
-  insuranceOptions,
-  insuranceSnapshotUploadAttributes,
-} from "../../../utils/insuranceDetailsUtils";
 
+const insuranceNumberInputAttributes = {
+  name: ["insuranceDetails", "insuranceNumber"],
+  label: "Insurance Number",
+  rules: [{ required: true, message: "Please enter your insurance number!" }],
+};
+const insuranceCompanySelectAttributes = {
+  name: ["insuranceDetails", "insuranceCompany"],
+  label: "Insurance Company",
+  rules: [{ required: true, message: "Please select insurance company!" }],
+};
+const insuranceSnapshotUploadAttributes = {
+  name: ["insuranceDetails", "insuranceSnapshot"],
+  label: "Insurance Document Snapshots",
+  rules: [{ required: true, message: "Please upload insurance snapshots!" }],
+};
+const insuranceOptions = [
+  { label: "Yes", value: "yes" },
+  { label: "No", value: "no" },
+];
 export default function InsuranceDetails({ form }) {
   const [haveInsurance, setHaveInsurance] = useState("");
   const handleHaveInsuranceChange = (value) => {
