--- src/app/utils/userInformationFormUtils.js
@@ -1,4 +1,0 @@
-export const layout = {
-  labelCol: { span: 6 },
-  wrapperCol: { span: 16 },
-};
